Add Mockups
This PR adds mockups for the initial focus, which is the white piece of paper placed on WP-admin which in which you write the _content_. We will get to the surrounding _metadata_ as soon as possible.

As mockups get updated, I will endeavor to keep the filenames the same, so we can compare using GitHub's wonderful binary difftool.