Change responsive padding to behave like max-width.
This commit changes the behavior of the padding by using calc. The net result is a behavior that is almost the same as if the contents had a max-width.

However it's hacky, and further complicated by the fixed positioning. I'm not sure whether this is good enough to merge, but I'm pushing so as to start the discussion at least.