Polish paddings, margins, headings (#651)
* Darken mover.

* Rearrange CSS, and lighten save state.

This commit moves some CSS from the button component to the tools stylesheet. It's a better place.

It also lightens the save state so it's the same color as everything else in the editor bar.

* Normalize block boundary paddings.

There were some offset here and there. This commit makes them 1:1 with the mockups, and also moves to using variables and a little SCSS math that should hopefully make the resulting positioning stuff a little easier to work with.

* Fix regression that hid the first blocks toolbar.

* Polish headings

This tunes the margins a bit. It also resizes headings 1-6 a bit, so they are differentiated and all different. This fixes #625

* Use padding instead of margin for heading.

* Move to tabs.

* Relabel a margin variable.