Cleaning up tests.
Cleans up the tests to make use of spies.  This also changes the
`props()` convention to check whether the keys are included in the
object.  include is used instead of equal to ensure that the tests do
not look at all props like children. Adds in a correction to the function
signature of Toolbar so that the component renders properly.