Parser: Store attributes in unescaped JSON format (#1213)
* Parser: Store attributes in unescaped JSON format

Previously we borrowed the `name="value"` format of tag attributes from
HTML. This caused some headache with blocks because of the different
ways we are wanting to save data in the block comment headers, not the
least of which reason is due to how data must or need not be escaped.

In this change we overhaul the format of the data to allow for (mostly)
unescaped JSON data in the block comment headers. Minimal escaping
exists to prevent breaking the HTML comment and to prevent
poorly-implemented tools from messing up the comments as well.