Blocks: Verify node exists before calling focus
It doesn't necessarily follow that this should be necessary, since refs should be assigned during componentDidMount. Specifically in the case that the error boundary catches an error though, the ref is not assigned. This may be related to how React, by its error message, "recreate[s] this component tree from scratch". Do we need a separate component for error boundary rendered as a wrapper of the block edit function? Might interfere with styling.