Try sliding toolbar for overflow on mobile
This PR may fix #2287, at least partially, or in a transitional way. It's a two-line code change, and all it does is ensure that on mobile, overflow can be scrolled.

This is not a good solution for desktop, of course, where the scrollbar when visible (as it is on nearly all computers) would not only cover content and be ugly, but scrolling horizontally is generally not a good experience. However on mobile, the scrollbar fades in briefly when you load the page, and you can just drag with your finger.

Let's discuss whether this is a good solution.