Hide Extended Settings unless Sidebar is opened
This fixes #2849, props @DevinWalker for concept.

Extended settings is the panel at the bottom that holds extra metaboxes. It is present when a plugin adds such metaboxes, otherwise it's not there. Given that the purpose of being able to dismiss the sidebar is for a simpler, lighter, writing experience, it makes sense to hide this panel if the sidebar panel is hidden. This PR does that.