Reduce the visual weight of blocks (#3068)
This implements an alternate block style that shows borders only during selection—on hover, just the mover and settings are shown with a small border. It also simplifies the appearance of multi-selection to match regular selection within text.

Finally, it consolidates keydown event handling to count arrow keys for invoking `isTyping` mode. This greatly reduces the feel of weight in the editor during writing.