Improve text columns focus style.
In the previous focus style commit, the focus style for these columns was a bit heavy handed.

This PR changes the focus style to mimic that of a selected text block. In the not too distant future, this block itself will likely be revisited with the recent nesting feature in mind, so this improves it until that time.