Reusable Blocks: Add reusable blocks effects (#3377)
* Add reusable blocks effects

Adds the effects necessary for supporting reusable blocks. There are 4:

- FETCH_REUSABLE_BLOCKS: Loads reusable blocks from the API and inserts
  them into the store.
- SAVE_REUSABLE_BLOCK: Persists a reusable block that's in the store to
  the API.
- MAKE_BLOCK_STATIC: Transforms a reusable block on the page into a
  regular block.
- MAKE_BLOCK_REUSABLE: Transforms a regular block on the page into a
  reusable block.

* Fix 'cannot call initializeMetaBoxes on undefined' error

`gutenberg_collect_meta_box_data` expected that
`window._wpGutenbergEditor` would always be instantiated, but this is
only true for when `wp.api.init()` returns quickly because of the API
schema being cached.

* Rename _wpGutenbergEditor -> _wpLoadGutenbergEditor

Make it clear that this variable is a promise by giving it a non-noun
name.