Pasting: Convert unknown shortcodes to Shortcode block
Blocks can specify shortcodes from which blocks should be automatically
created. For any other shortcode, the core Shortcode block will now act
as a fallback.

The main benefit of this is that, within the Shortcode block, the code
will be properly handled and never mishandled. Notably, this fixes the
case where a shortcode sitting in a Paragraph block will be picked up
with its double quotes rendered as `&doubl;`, thus breaking parsing of a
shortcode's attributes.