Split variables stylesheet into multiple
This PR splits out breakpoints and colors from the _variables.scss stylesheet, in a hope to fix #964.

It also uses `//` for comments that should be eaten in compilation.

Further improvements can probably be made if we find either the need for more variables in either of the current sections in _variables.scss, or if we can move some of those variables to live in context of the components that use them. As of now, though, the remaining variables don't feel like enough to split further, and the are used across many components.