Try isEditing focus
This PR is an experiment to add a focus style for blocks have focus when using just the keyboard. It is experimental right now, but it could potentially provide a nicer editing flow, by showing what's selected when your caret disappears and a block is instead selected. Obviously when that happens you can press delete or backspace to remove said block, or you can press enter to make a linebreak.