Fix issue where black rectangle would briefly blink on new paragraphs
In master, if you press enter a lot and type a lot, a black rectangle sometimes briefly appears. This is the isEditing rectangle we use to indicate where the cursor is, if it's not in text, i.e. if you have a placeholder selected.

This PR adds a teensy delay to this so it shouldn't show up.