Try fixing nav menu scroll
This fixes #4445. Props @m-e-h to starting the work to address this.

This removes `overflow: hidden;` from the `body` element, which was previously applied at the desktop breakpoint. That means if you have a very tall navigation menu, or your viewport isn't very tall, then you can at least scroll the main container to reveal all menu items.

This is not an ideal fix, as it adds another scrollbar in this edgecase. But given it's an edgecase, and given the benefits of preventing scroll bleed the individually scrolling main content area has, perhaps it's worth it?

We need to test this PR to see that there aren't other side effects.

We also tried letting the nav container itself scroll, but that crops the flyout menu.