Fixed quote to heading transformation. (#5088)
This commit corrects the logic of quote to heading transform.
The logic is now the following:
If no citation and no content exist an empty heading is created.
If citation exists but the content does not exist a heading is created with citation as the content.
In other citations the first paragraph of the quote is transformed into and heading and new quote is created equal to the original but with the first paragraph removed. If the quote just contains one paragraph the new quote will contain just the citation.