Fixed empty after images when the images are inside an anchor. (#5485)
This commit fixes a problem where empty paragraphs are created after images if the image is inside an anchor.
This happens because we have a logic to move images out of paragraphs so they are parsed. If the images were inside an anchor we move the images but left the anchor this lead to the creation of empty paragraphs with just the anchor.