Added an effect that solves a bug, where adding inserting a block right after removal, inserts a block in the top of the document. (#5568)
The problem was that block selection becomes invalid after block removal, was the selection was still pointing to the removed blocks UID's. Now we select the block that was before the removed block(s). So if we insert a block right after removing the inserted block goes to the position of the removed block.