Fixed: contrast checker display. (#5800)
Notices component changed and now instead of receiving a content prop renders the children. Contrast checker is now updated to take into account this recent change of the notices component.

The test cases of ContrastChecker did not fail because they tested that the notices are shown when needed using the notice component, but did not test if the notices component, in fact, rendered the message. The use of shallow was changed to mount in the tests, as I think mount in this case, gives a bigger assurance that messages are in fact rendered. It may create false fails, e.g: class changes in notices. But when doing this changes we should probably test ContrastChecker so this fails just increase the probability of the manual inspection happening so they are good in my option.