Fixed clone function to clone innerBlocks.
Clone function did not cloned the innerBlocks so if we clone a blocks e.g: a collumn with innerBlocks both collumns would reference the same innerBlocks so changed applied to one innerBlock would also apply to the other.
Now if innerBlocks are not passed explicitly to clone, all the innerBlocks are recursively cloned.