Add polish (#6138)
* Polish link dialog.

This uses variables for colors and shadows, and also fixes a regression with the link dialog when linking an image.

However there's still a bug here, where as soon as you start typing in the image link dialog, the entire thing disappears and you can't type. This is possibly due to isEditing mode being invoked there. Can you take a look, @noisysocks @karmatosed?

* Unify markup between More block and Pagination block

This adds a white background, and some things, to the Pagination block. This will only be visible if a user loads a stylesheet into the editor.

Fixes https://github.com/WordPress/gutenberg/pull/1467#discussion_r180033504, cc @swissspidy.

* Unify button sizes, and slim down "Media Library".

This makes the placeholders fit better in small breakpoints.

* Change left arrow on breadcrumbs.

This makes the left arrow consistent with the one used when creating links on images.