Fixed image add link input disappears issue (#5942)
* #5894:Image add link input disappear when start typing, fixed it by stopping event propagation

Signed-off-by: Faishal Saiyed <saiyedfaishal@gmail.com>

* Fix issues with suggestions not working in image block

The issue was that on mobile, the block toolbar has overflow auto so as to scroll horizontally. This hid the suggestions box in a weird scrollbar.

This PR tweaks that, so the suggestions don't show up on mobile at all. This is not an ideal fix, but suggestions weren't very mobile friendly in the first place.

Suggest we get this in now as a bandaid, and then look at addressing suggestions on mobile separately, for example using the same technique the block switcher uses.

* Fix issue with caption links

* Ignore typing in block toolbars