Refactor blocks->PanelColor, to avoid the need for the colorName prop (#6543)
With the creation of a PanelColor in blocks, there is no need to receive the colorName anymore as the component already has the color context available and can get the colorName from it.
This change makes the usage of the component easier. Existing core blocks that make use of the component were updated.