Fix issue with high contrast indicator in Edge (#6373)
This rejiggers the Switch high contrast mode a bit, to enhance how it works in Edge. Edge has an obscure bug where it overlays the outline on the SVG, regardless of outline-offset. This fixes that, and makes the switch work in high contrast mode in IE, Edge and Firefox on Windows. Chrome on Windows does not support high contrast mode.

The switch still works unchanged on Chrome and Firefox MacOS.