Try fixing captions on resized images (#6496)
* Try fixing captions on resized images

This is an alternative, and probably better approach, than the one explored in https://github.com/WordPress/gutenberg/pull/6304.

Both branches try and fix the same issue — ensuring that the caption inside a `figure` is only ever as wide as the adjecent image. This is only ah issue when that image is resized.

This branch fixes that, in a fairly clean way. See the details in https://css-tricks.com/almanac/properties/w/width/

* Add fix for Edge and IE11

* Polish up Edge min-content trick.

* Add deprecated version.

* Fix deprecated version