Improve child blocks API's and UI. (#7003)
This PR extends the blocks API with two new selectors getChildBlockNames and hasChildBlocks.
A class is added to inserter items with children, and the UI is changed using the new class. Now we have a special design in blocks with children.
A new area that appears when we are inserting blocks inside blocks with children was created.