Support opt-in visual styles for core blocks
The purpose of this change is to offer default visual styles for themes
that want them without interfering with existing themes that provide
their own. Core blocks may include a `theme.scss` file containing visual
styles, and those styles will be built into
`build/core-blocks/theme.css`. The visual styles are enqueued for
editing but are not enqueued for viewing unless there is theme support
for `wp-block-styles`.