Try a new hover and select approach to improve nested block selection (#6773)
* Alternate hover approach WIP

* Show outline on parent blocks

* Block Breadcrumb: Render arrow as RTL compatible

* Writing Flow: Force tab stop at focusable wrapper when inner blocks exist

* Block List: Remove unused variables

* Inner Blocks: Show overlay for small screen, unselected root block

* Try a refined approach to drag and drop

This commit changes how drag and drop works, so that nested blocks do not have any invisible side affordance for the grabby handle. You can still use the inner padding for dragging, and in a separate enhancement we should make the hover label draggable.

This makes it easier to select the parent block, by allowing you to click the available space below the side UI to select the parent.

Kinda.

There is still an affordance for hovering the side of the block to access the side UI (movers, settings, trash), and this extends the height of 4 stacked buttons on the side. If you hover the area below that, yes you can click the thing below.

This is a balancing act, and I'd like for you to try this. The branch has been thoroughly rebased and squashed prior to this to make it easy to undo this change if it doesn't feel right.

* Fix the empty appender in more than columns block

The empty appender, when on mobile, or in a nested block, shows a reduced version. Previously this was only in the columns block, now it's any block that uses nested children.

* Tune hover label again.

* Fix rebase issue.

* hide outlines from empty default blocks

* Fix paddings and hover breadcrumb on classic block.

* Fade the title block border.

* Address feedback.

Add border to switcher component instead.

Add comment to explain z index.