Fix issue with normal buttons with icons (#7073)
* Fix issue with normal buttons with icons

Reported in https://github.com/WordPress/gutenberg/pull/5350#issuecomment-393621436, this PR fixes an issue where the icon was misaligned when using it with a standard Button look.

This scopes and adds some CSS to fix and align things. If this isn't the right approach, another approach would be to use the Button component here instead of the IconButton component (which has that no-background look). But then we would have to allow passing an icon to the Button component.

* Fix tab accident!

* Address feedback.

Inverse the behavior for excluding normal buttons.