New User Experience (NUX) (#6631)
* Fix appearance of DotTip on mobile

Force DotTip content to appear underneath the indicator dot when on
mobile viewports. This requires overriding a lot of .component-popover
styles which is quite gross.

* Remove unnecessary <Fragment>

* Position NUX dots right next to the edge of the button

Adjust popover's layout algorithm so that if yAxis === 'middle', it will
set popoverLeft such that the popover points to the edge of the anchor
node. This aligns with what e.g. position="bottom center" does.

* Increase NUX tip padding

* Add margin to sides of NUX tip on mobile

* Don't encourage using domReady in the NUX README

* Add wp-nux as a dependency of wp-edit-post

* Suggest that NUX tip identifiers are prefixed

* Don't use partial(), which is slated for removal

* tipID → tipId

* Improve NUX JSDoc comments

* Memoize the getAssociatedGuide() selector

* NUX: Dismiss the tip when one clicks away from it

* Position tooltips above NUX tips

* Remove aria-modal property from DotTip

Safari 11 has a weird bug when this property is used.

* Make order of the markup in DotTip match how they appear visually

* Improve a11y of DotTip labels

* Implement New User Guide

Adds a series of floating modal 'tips' which introduces a new user to
the editor. These tips can be advanced one by one or dismissed
alltogether. If dismissed, they will never show again.

* Dismiss tips instead of disabling when the X is clicked

* Fix DotTip focus issue

The first DotTip should receive focus when the page loads, and focus
should not be on the X button.

Also improve the note explaining our temporary position workaround.

* NUX: Only show inserter tip in DefaultBlockAppender

Only display the first DotTip in the new user guide in the
DefaultBlockAppender that appears in a new post. This prevents the guide
from appearing in an existing post that contains empty blocks.