Fix IE11 formatting toolbar visibility (#7413)
IE11 does not support `position: sticky`, and we were using it
unconditionally to position the formatting toolbar, resulting in
a toolbar that was only partially visible. This updates our
styles to start with `position: relative` which also provides
a coordinate system but does not have the sticky behavior. Then
we apply `position: sticky` for the browsers that support it.