Move the remove icon button to the ellipsis menu as a menu button (#7543)
* Move BlockRemoveButton into NavigableMenu of BlockSettingsMenu

* Update BlockSettingsMenu BlockRemoveButton appearance

Add the label 'Remove Block' to the BlockRemoveButton and change the classname
to editor-block-settings-menu__control so the the icon button is consistent
with other menu controls

* Add unit tests for BlockRemoveButton

* Extract button for deleting shared block into its own component

Extract the button into a new SharedBlockDeleteButton, so that it can be
rendered separately and so that a separator can be added above it.

* Extract test for SharedBlockDeleteButton from SharedBlockSettings

Refactor tests to match refactoring to components. Add a snapshot test
for SharedBlockDeleteButton

* Add a div that has a separator class applied to the BlockSettingsMenu

* Rename SharedBlockSettings component to SharedBlockConvertButton

Since a recent refactor, this component now only renders buttons for
converting to and from shared blocks. This commit renames it to that
effect.